Submission for Assessment #1 CODE1240
A video of my submission for Assessment one, a performance space based
on the song C.U.B.A by Calvin Harris.

The Video shows the three spaces that were designed to correspond to the
three distinct parts of the song, the immersion, the ascension and the
silence. Each room is designed to give the listen/viewer a different
experiential experience.